% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.binomial}
\alias{glm.binomial}
\title{Creates a Binomial GLM family object.}
\usage{
glm.binomial(y, weights = NULL, link = "logit")
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}

\item{link}{The link function type.}
}
\description{
Creates a Binomial GLM family object.
}
