% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.poisson}
\alias{glm.poisson}
\title{Creates a Poisson GLM family object.}
\usage{
glm.poisson(y, weights = NULL)
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}
}
\description{
Creates a Poisson GLM family object.
}
