% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_elnet.R
\name{group_basil}
\alias{group_basil}
\title{Group basil function}
\usage{
group_basil(
  X,
  y,
  groups,
  group_sizes,
  alpha = 1,
  penalty = sqrt(group_sizes),
  user_lmdas = double(0),
  max_n_lambdas = 100,
  n_lambdas_iter = 5,
  use_strong_rule = TRUE,
  do_early_exit = TRUE,
  verbose_diagnostic = FALSE,
  delta_strong_size = 5,
  max_strong_size = ncol(X),
  max_n_cds = as.integer(1e+05),
  tol = 1e-07,
  rsq_slope_tol = 0.001,
  rsq_curv_tol = 0.001,
  newton_tol = 1e-08,
  newton_max_iters = 100,
  min_ratio = 0.01,
  n_threads = 16
)
}
\description{
Group basil function
}
