% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.snp_phased_ancestry}
\alias{io.snp_phased_ancestry}
\title{IO handler for SNP phased, ancestry matrix.}
\usage{
io.snp_phased_ancestry(filename, read_mode = "file")
}
\arguments{
\item{filename}{File name.}

\item{read_mode}{Reading mode.}
}
\description{
IO handler for SNP phased, ancestry matrix.
}
