% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.dense}
\alias{matrix.dense}
\title{Creates a viewer of a dense matrix.}
\usage{
matrix.dense(mat, method = "naive", n_threads = 1)
}
\arguments{
\item{mat}{The dense matrix.}

\item{method}{Method type.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a viewer of a dense matrix.
}
