% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.one_hot}
\alias{matrix.one_hot}
\title{Creates a one-hot encoded matrix.}
\usage{
matrix.one_hot(mat, levels = NULL, n_threads = 1)
}
\arguments{
\item{mat}{The dense matrix.}

\item{levels}{Levels.}

\item{n_threads}{Number of threads.}
}
\value{
One-hot encoded matrix.
}
\description{
Creates a one-hot encoded matrix.
}
\examples{
n <- 100
p <- 20
mat <- matrix(rnorm(n * p), n, p)
out <- matrix.one_hot(mat)
}
