% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.snp_phased_ancestry}
\alias{matrix.snp_phased_ancestry}
\title{Creates a SNP phased, ancestry matrix.}
\usage{
matrix.snp_phased_ancestry(io, n_threads = 1)
}
\arguments{
\item{io}{IO handler.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a SNP phased, ancestry matrix.
}
