% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.snp_unphased}
\alias{matrix.snp_unphased}
\title{Creates a SNP unphased matrix.}
\usage{
matrix.snp_unphased(io, n_threads = 1)
}
\arguments{
\item{io}{IO handler.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a SNP unphased matrix.
}
