% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.standardize}
\alias{matrix.standardize}
\title{Creates a standardized matrix.}
\usage{
matrix.standardize(mat, centers = NULL, scales = NULL, ddof = 0, n_threads = 1)
}
\arguments{
\item{mat}{The underlying matrix.}

\item{centers}{The center values.}

\item{scales}{The scale values.}

\item{ddof}{Degrees of freedom.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a standardized matrix.
}
