% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.subset}
\alias{matrix.subset}
\title{Creates a subset of the matrix along an axis.}
\usage{
matrix.subset(mat, indices, axis = 0, n_threads = 1)
}
\arguments{
\item{mat}{The matrix to subset.}

\item{indices}{Array of indices to subset the matrix.}

\item{axis}{The axis along which to subset.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a subset of the matrix along an axis.
}
