% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{set_configs}
\alias{set_configs}
\title{Set configuration settings.}
\usage{
set_configs(name, value = NULL)
}
\arguments{
\item{name}{Configuration variable name.}

\item{value}{Value to assign to the configuration variable.}
}
\description{
Set configuration settings.
}
