% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint.box}
\alias{constraint.box}
\title{Creates a box constraint.}
\usage{
constraint.box(lower, upper, max_iters = 100, tol = 1e-09)
}
\arguments{
\item{lower}{Lower bound.}

\item{upper}{Upper bound.}

\item{max_iters}{Maximum number of proximal Newton iterations.}

\item{tol}{Convergence tolerance for proximal Newton.}

\item{n_threads}{Number of threads.}
}
\value{
Box constraint.
}
\description{
Creates a box constraint.
}
\examples{

lower <- double(10)
upper <- lower + 5
c <- constraint.box(
    lower=lower,
    upper=upper
)
}
\author{
Trevor Hastie and James Yang\cr Maintainer: Trevor Hastie \href{mailto:hastie@stanford.edu}{hastie@stanford.edu}
}
