% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.binomial}
\alias{glm.binomial}
\title{Creates a Binomial GLM family object.}
\usage{
glm.binomial(y, weights = NULL, link = "logit")
}
\arguments{
\item{y}{Binary response vector, with values 0 or 1, or a logical vector.}

\item{weights}{Observation weight vector, with default \code{NULL}.}

\item{link}{The link function type, with choice \code{"logit"} (default) or \code{"probit"}).}
}
\value{
Binomial GLM object.
}
\description{
Creates a Binomial GLM family object.
}
\examples{
n <- 100
y <- rbinom(n, 1, 0.5)
obj <- glm.binomial(y)
}
