% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.cox}
\alias{glm.cox}
\title{Creates a Cox GLM family object.}
\usage{
glm.cox(
  start = -Inf,
  stop,
  status,
  weights = NULL,
  tie_method = c("efron", "breslow")
)
}
\arguments{
\item{start}{Start time vector. Default is a vector of \code{-Inf} of same length as \code{stop}.}

\item{stop}{Stop time vector.}

\item{status}{Binary status vector of same length ast \code{stop}, with 1 a "death", and 0 censored.}

\item{weights}{Observation weights, with default \code{NULL}.}

\item{tie_method}{The tie-breaking method - one of  \code{"efron"} (default) or \code{"breslow"}.}
}
\value{
Cox GLM object.
}
\description{
Creates a Cox GLM family object.
}
\examples{
n <- 100
start <- sample.int(20, size=n, replace=TRUE)
stop <- start + 1 + sample.int(5, size=n, replace=TRUE)
status <- rbinom(n, 1, 0.5)
obj <- glm.cox(start, stop, status)
}
