% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.gaussian}
\alias{glm.gaussian}
\title{Creates a Gaussian GLM family object.}
\usage{
glm.gaussian(y, weights = NULL, opt = TRUE)
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}

\item{opt}{If \code{TRUE}, an optimized routine is run.}
}
\value{
Gaussian GLM
}
\description{
Creates a Gaussian GLM family object.
}
\examples{
n <- 100
y <- rnorm(n)
obj <- glm.gaussian(y)
}
