% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.multigaussian}
\alias{glm.multigaussian}
\title{Creates a MultiGaussian GLM family object.}
\usage{
glm.multigaussian(y, weights = NULL, opt = TRUE)
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}

\item{opt}{If \code{TRUE}, an optimized routine is run.}
}
\value{
MultiGaussian GLM object.
}
\description{
Creates a MultiGaussian GLM family object.
}
\examples{
n <- 100
K <- 5
y <- matrix(rnorm(n*K), n, K)
obj <- glm.multigaussian(y)
}
