% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.multinomial}
\alias{glm.multinomial}
\title{Creates a Multinomial GLM family object.}
\usage{
glm.multinomial(y, weights = NULL)
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}
}
\value{
Multinomial GLM object.
}
\description{
Creates a Multinomial GLM family object.
}
\examples{
n <- 100
K <- 5
y <- t(rmultinom(n, 1, rep(1/K, K)))
obj <- glm.multinomial(y)
}
