% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.multinomial}
\alias{glm.multinomial}
\title{Creates a Multinomial GLM family object.}
\usage{
glm.multinomial(y, weights = NULL)
}
\arguments{
\item{y}{Response vector.}

\item{weights}{Observation weights.}
}
\description{
Creates a Multinomial GLM family object.
}
