% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{io.snp_unphased}
\alias{io.snp_unphased}
\title{IO handler for SNP unphased matrix.}
\usage{
io.snp_unphased(filename, read_mode = "file")
}
\arguments{
\item{filename}{File name.}

\item{read_mode}{Reading mode.}
}
\description{
IO handler for SNP unphased matrix.
}
