% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.block_diag}
\alias{matrix.block_diag}
\title{Creates a block-diagonal matrix.}
\usage{
matrix.block_diag(mats, n_threads = 1)
}
\arguments{
\item{mats}{List of matrices.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a block-diagonal matrix.
}
