% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.interaction}
\alias{matrix.interaction}
\title{Creates a matrix with pairwise interactions.}
\usage{
matrix.interaction(mat, intr_keys, intr_values, levels = NULL, n_threads = 1)
}
\arguments{
\item{mat}{The dense matrix.}

\item{intr_keys}{List of feature indices.}

\item{intr_values}{List of list of feature indices.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a matrix with pairwise interactions.
}
