% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.lazy_cov}
\alias{matrix.lazy_cov}
\title{Creates a lazy covariance matrix.}
\usage{
matrix.lazy_cov(mat, n_threads = 1)
}
\arguments{
\item{mat}{The data matrix.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a lazy covariance matrix.
}
