% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix.sparse}
\alias{matrix.sparse}
\title{Creates a viewer of a sparse matrix.}
\usage{
matrix.sparse(mat, method = "naive", n_threads = 1)
}
\arguments{
\item{mat}{The sparse matrix to view.}

\item{method}{Method type.}

\item{n_threads}{Number of threads.}
}
\description{
Creates a viewer of a sparse matrix.
}
